cleanup and divide dashboard table metadata
